Change speed field type from Float to Integer in Teleprompter model

+++ src/main/java/juancarlos/tfg/teleprompter/models/Teleprompter.java
@@ -39,7 +39,7 @@ public class Teleprompter {
     @JoinColumn(name = "user_id")
     private User user;
 
-    private Float speed;
+    private Integer speed;
 
     private Integer type;
